fix: 修復K線圖和即時報價問題
主要修復:
1. K線圖欄位映射問題
   - 修復 FinMind API 欄位名稱不匹配 (max/min → high/low, Trading_Volume → volume)
   - 添加正確的類型轉換 (float, int, ISO 日期字串)
   - 位置: backend/app/services/stock_data_service.py

2. 即時報價延遲問題
   - 整合 TWSE 官方 API 作為主要即時數據來源
   - 實現優先順序: TWSE → Fugle → FinMind
   - 股價延遲從 4+ 小時降至 5-20 秒
   - 新增: backend/app/data_fetchers/twse_fetcher.py::get_realtime_quote()
   - 修改: backend/app/services/stock_data_service.py::get_realtime_price()

3. 前端錯誤處理改進
   - 顯示詳細的錯誤類型和狀態碼
   - 添加可能原因提示
   - 改進重試按鈕樣式
   - 位置: frontend/lib/screens/stock_detail_screen.dart

4. API 測試工具
   - 新增 backend/app/test_api.py (後端測試腳本)
   - 新增 test_web.html (前端測試頁面)

測試結果:
- ✅ K線圖數據載入成功 (41筆歷史數據)
- ✅ 所有欄位正確 (date, open, high, low, close, volume)
- ✅ 即時報價顯示最新數據 (2330 台積電: 1775.0)

+++ taiwan_stock_app/backend/app/schemas/watchlist.py
@@ -15,8 +15,8 @@ class WatchlistAdd(BaseModel):
 class WatchlistItem(BaseModel):
     stock_id: str
     name: str
-    current_price: Decimal
-    change_percent: Decimal
+    current_price: float
+    change_percent: float
     added_at: datetime
     notes: Optional[str] = None
